+++ tests/Kiota.Builder.Tests/Writers/Php/CodeMethodWriterTests.cs
@@ -287,11 +287,10 @@ public class CodeMethodWriterTests : IDisposable
 
         Assert.Contains("public function post(): Promise", result);
         Assert.Contains("$requestInfo = $this->createPostRequestInformation();", result);
-        Assert.Contains("RejectedPromise", result);
         Assert.Contains("@link https://learn.microsoft.com/ Learning", result);
-        Assert.Contains("catch(Exception $ex)", result);
         Assert.Contains("'401' => [Error401::class, 'createFromDiscriminatorValue']", result);
-        Assert.Contains("return $this->requestAdapter->sendPrimitiveAsync($requestInfo, StreamInterface::class, $errorMappings);", result);
+        Assert.Contains("$result = $this->requestAdapter->sendPrimitiveAsync($requestInfo, StreamInterface::class, $errorMappings);", result);
+        Assert.Contains("return $result;", result);
     }
 
     [Fact]
@@ -380,13 +379,14 @@ public class CodeMethodWriterTests : IDisposable
         _codeMethodWriter.WriteCodeElement(codeMethod, languageWriter);
         var result = stringWriter.ToString();
 
+        Assert.Contains("@throws Exception", result);
         Assert.Contains("public function post(): Promise", result);
         Assert.Contains("$requestInfo = $this->createPostRequestInformation();", result);
-        Assert.Contains("RejectedPromise", result);
         Assert.Contains("@link https://learn.microsoft.com/ Learning", result);
-        Assert.Contains("catch(Exception $ex)", result);
         Assert.Contains("'401' => [Error401::class, 'createFromDiscriminatorValue']", result);
-        Assert.Contains("return $this->requestAdapter->sendPrimitiveAsync($requestInfo, PhoneNumberPrefix::class, $errorMappings);", result);
+        Assert.Contains("/** @var Promise<PhoneNumberPrefix|null> $result", result);
+        Assert.Contains("$result = $this->requestAdapter->sendPrimitiveAsync($requestInfo, PhoneNumberPrefix::class, $errorMappings);", result);
+        Assert.Contains("return $result;", result);
     }
 
     public static IEnumerable<object[]> SerializerProperties => new List<object[]>